feat(linker): consolidator user testing feedback
* Forkflow changed - when user tries to send message to user support, new dialog will be opened. In
this dialog is textarea with general message which can be changed by the user. This message will be merged with some technical info before sending to user support.
* Some texts were too.
