+++ apps/linker/src/styles.scss
@@ -185,3 +185,8 @@ button:focus {
 mat-icon {
   overflow: inherit !important;
 }
+
+.page-subtitle {
+  margin-bottom: 1rem;
+  font-size: 1.5rem;
+}
